chore(liveness): move error types to separate file (#3998)
