+++ packages/react-liveness/src/components/FaceLivenessDetector/service/types/index.ts
@@ -1,3 +1,4 @@
 export * from './faceDetection';
 export * from './liveness';
 export * from './machine';
+export * from './error';
